+++ src/iotvm-eventengine/src/main/java/org/softwareforce/iotvm/eventengine/cep/ct/AverageCalculationCompositeTransformationParameters.java
@@ -178,6 +178,7 @@ public class AverageCalculationCompositeTransformationParameters
 
   /** TODO EXPERIMENTAL. */
   public int getFutureWindowsLookupAlternative() {
+    // TODO Use only this!
     if (this.futureWindowsLookup > 0 && this.pastWindowsLookup > 0) {
       return this.futureWindowsLookup + this.pastWindowsLookup;
     }