+++ src/iotvm-extensions/iotvm_extensions/examples/average_calculation_parameters_sets.py
@@ -46,6 +46,7 @@ def generate_composite_transformation_parameters_ids(
     multiplier: int = 1
 
     for combination in combinations:
+        # TODO S is wrong!
         ctp_id: str = (
             f"{_PREFIX}{combination[0].lower()}_PT{combination[1]}S_null_null_{combination[2]}_{str(combination[3]).lower()}_{combination[4]}_PT{combination[1] * multiplier}S_{combination[5]}"
         )
@@ -56,17 +57,19 @@ def generate_composite_transformation_parameters_ids(
 
 @dataclass
 class CompositeTransformationParameterID:
-    # TODO physical quantity.
-    time_window_size: int
+    physical_quantity: str  # TODO Rename
+    time_window_size: str
     number_of_contributing_sensors: int
-    ignore_completeness_filtering: bool
-    fabrication_past_events_steps_behind: int
-    fabrication_forecasting_time_window_size: int
-    fabrication_forecasting_steps_ahead: int
+    ignore_completeness_filtering: bool  # TODO Remove.
+    fabrication_past_events_steps_behind: int  # TODO Rename
+    fabrication_forecasting_time_window_size: str  # TODO Remove
+    fabrication_forecasting_steps_ahead: int  # TODO Reename
 
     def __str__(self) -> str:
         string: str = (
-            f"time window size of {self.time_window_size} secs with at least {self.number_of_contributing_sensors} sensors"
+            f"physical quantity {self.physical_quantity}, "
+            f"time window size of {self.time_window_size} secs "
+            f"with at least {self.number_of_contributing_sensors} sensors"
         )
 
         if self.fabrication_past_events_steps_behind == 0 and self.fabrication_forecasting_steps_ahead == 0:
@@ -88,34 +91,31 @@ class CompositeTransformationParameterID:
 
 
 def parse_ctp_id(ctp_id: str) -> CompositeTransformationParameterID:
-    # TODO Fix.
-    ctp_id = ctp_id.removeprefix("w_avg_temperature_")
-    # TODO Physical quantity!
+    ctp_id = ctp_id.removeprefix(_PREFIX)
     parts: List[str] = ctp_id.split("_")
 
-    time_window_size: str = parts[0]
-    time_window_size = time_window_size.removeprefix("PT")
-    time_window_size = time_window_size.removesuffix("S")
+    physical_quantity: str = parts[0]
 
-    number_of_contributing_sensors: str = parts[3]
+    time_window_size: str = parts[1]
 
-    ignore_completeness_filtering: str = parts[4]
+    number_of_contributing_sensors: str = parts[4]
+
+    ignore_completeness_filtering: str = parts[5]
     assert ignore_completeness_filtering in ["true", "false"]
 
-    fabrication_past_events_steps_behind: str = parts[5]
+    fabrication_past_events_steps_behind: str = parts[6]
 
-    fabrication_forecasting_time_window_size: str = parts[6]
-    fabrication_forecasting_time_window_size = fabrication_forecasting_time_window_size.removeprefix("PT")
-    fabrication_forecasting_time_window_size = fabrication_forecasting_time_window_size.removesuffix("S")
+    fabrication_forecasting_time_window_size: str = parts[7]
 
-    fabrication_forecasting_steps_ahead: str = parts[7]
+    fabrication_forecasting_steps_ahead: str = parts[8]
 
     return CompositeTransformationParameterID(
-        time_window_size=int(time_window_size),
+        physical_quantity=physical_quantity,
+        time_window_size=time_window_size,
         number_of_contributing_sensors=int(number_of_contributing_sensors),
         ignore_completeness_filtering=ignore_completeness_filtering == "true",
         fabrication_past_events_steps_behind=int(fabrication_past_events_steps_behind),
-        fabrication_forecasting_time_window_size=int(fabrication_forecasting_time_window_size),
+        fabrication_forecasting_time_window_size=fabrication_forecasting_time_window_size,
         fabrication_forecasting_steps_ahead=int(fabrication_forecasting_steps_ahead),
     )